requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0
cryptography==42.0.5
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
jinja2==3.1.3
